v2.0.0-rc2: Merge stash fixes (softmax, conv2d groups, ONNX input_shapes)
Merge stashed@{0} from feat/v2-compiler-rewrite branch:

Softmax fix:
  - Change softmax dim from usize to i64 (negative dim indexing support)
  - Fix non-last-axis softmax: add stride-based access pattern instead of
    assuming contiguous rows (e.g. axis=2 on NCHW was broken)
  - Pass stride as second param to softmax dispatch kernel
  - Update API: Tensor::softmax, GraphBuilder::softmax, ONNX converter

Conv2d grouped convolution fix:
  - Fix weight validation: check input_ch == weight_ch * groups (not just
    input_ch == weight_ch, which broke grouped convolutions)
  - Replace fragile brute-force dim inference loop in f32 dispatch with
    explicit spatial dims passed as compile-time params
  - Fix bias expansion in nn/conv.rs: manually expand to spatial dims
    to avoid broadcasting issues
  - Update quantized conv2d dispatch (conv2d_u4/u8) to use the same
    9-param format with explicit spatial dims, eliminating its own
    fragile dim inference loop

ONNX input_shapes:
  - Add OnnxConverter::with_input_shapes() builder method
  - Python AotExecutor(input_shapes={'x': [-1,3,224,224]}) for dynamic inputs
  - Negative dim values become DimExpr::Symbol (e.g. d1 for -1)

Better shape inference errors:
  - Broadcast mismatch errors now include node id, opcode, and both shapes

Merge conflict resolved: both input_shapes and quantize params coexist.

+++ src/ir/builder.rs
@@ -770,7 +770,7 @@ impl GraphBuilder {
     }
 
     /// Softmax along a dimension.
-    pub fn softmax(&self, input: &GraphTensor, dim: usize) -> GraphTensor {
+    pub fn softmax(&self, input: &GraphTensor, dim: i64) -> GraphTensor {
         let output_type = input.tensor_type.clone();
         let mut attrs = HashMap::new();
         attrs.insert("axis".to_string(), dim.to_string());
